refactor[#235] : 맵 페이지에 있는 바텀 시트 컴포넌트 생성
