+++ src/app/FooterNavigationBar.tsx
@@ -1,11 +1,4 @@
-import { useRef } from "react";
-import { Sheet, SheetRef } from "react-modal-sheet";
-import { NavLink, useLocation } from "react-router-dom";
-import { useResearchMarkingList } from "@/features/map/hooks";
-import { useMapStore } from "@/features/map/store";
-import { MarkingItem } from "@/features/marking/ui";
-import { useGetMarkingList } from "@/entities/marking/api";
-import { MarkingList } from "@/entities/marking/ui";
+import { NavLink } from "react-router-dom";
 import { useGetProfile } from "@/entities/profile/api";
 import { API_BASE_URL, ROUTER_PATH } from "@/shared/constants";
 import { useAuthStore } from "@/shared/store";
@@ -20,82 +13,8 @@ const footerNavigationBarStyles = {
 export const FooterNavigationBar = () => {
   const { active, inactive, base } = footerNavigationBarStyles;
 
-  const location = useLocation();
-
-  const ref = useRef<HTMLDivElement>(null);
-  const sheetRef = useRef<SheetRef>();
-
-  const snapPoints = [-50, 0.5, 116];
-  const initialSnap = snapPoints.length - 1;
-
-  const snapPointRef = useRef(initialSnap);
-  const snapTo = (i: number) => sheetRef.current?.snapTo(i);
-
-  const { bounds } = useResearchMarkingList();
-  const { data: markingList } = useGetMarkingList({
-    southWestLat: bounds?.southWest.lat,
-    southWestLng: bounds?.southWest.lng,
-    northEastLat: bounds?.northEast.lat,
-    northEastLng: bounds?.northEast.lng,
-    sortType: "RECENT",
-  });
-
-  const mapMode = useMapStore((state) => state.mode);
-
   return (
-    <footer ref={ref} className="relative">
-      <Sheet
-        ref={sheetRef}
-        isOpen={location.pathname === "/map" && mapMode === "view"}
-        onClose={() => {
-          const isSheetTop = snapPointRef.current === 0;
-
-          if (isSheetTop) {
-            snapTo(snapPointRef.current - 1);
-            return;
-          }
-
-          snapTo(initialSnap);
-        }}
-        snapPoints={snapPoints}
-        initialSnap={initialSnap}
-        onSnap={(snapPointIndex) => (snapPointRef.current = snapPointIndex)}
-        style={{ zIndex: 1 }}
-        mountPoint={ref.current || document.querySelector("#root")!}
-      >
-        <Sheet.Container>
-          <Sheet.Header />
-          <Sheet.Content
-            style={{
-              padding: 0,
-              paddingBottom: sheetRef.current?.y,
-            }}
-          >
-            <Sheet.Scroller
-              draggableAt="both"
-              style={{
-                height: "calc(100% - 5rem)",
-              }}
-            >
-              {/* todo 버튼 활성화 여부에 따라 내용 바뀜 */}
-              <h1 className="title-1 text-grey-900 p-4">주변 마킹</h1>
-
-              <div className="px-4">
-                <MarkingList>
-                  {markingList?.map((marking) => (
-                    <MarkingItem
-                      key={marking.markingId}
-                      {...marking}
-                      onRegionClick={() => {}}
-                    />
-                  ))}
-                </MarkingList>
-              </div>
-            </Sheet.Scroller>
-          </Sheet.Content>
-        </Sheet.Container>
-      </Sheet>
-
+    <footer className="relative">
       <nav className="relative z-10">
         <ul className="flex justify-between items-center gap-2 bg-grey-0 px-2 h-20">
           <li className="grow">
