+++ src/widgets/map/ui/index.ts
@@ -3,3 +3,4 @@ export * from "./googleMaps";
 export * from "./mapControlWidget";
 export * from "./mapMarkerWidget";
 export * from "./mapInitializer";
+export * from "./mapBottomSheet";
